feat(ls): move configuration into app settings
